+++ pages/settings.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * The main Calendario Settings page.
+ *
+ * @package calendario
+ */
+
+?>
+
+<div class="wrap">
+	<h2>Editorial Calendar.io Settings</h2>
+</div>